got edit working to MVP standards. Finished Chat MVP.
