+++ src/scripts/chat/apiManager.js
@@ -1,3 +1,5 @@
+import printToDOM from "./printToDOM"
+
 const apiManager = {
   getUsername: () => {
     const activeUserId = sessionStorage.getItem("activeUser");
@@ -28,6 +30,9 @@ const apiManager = {
         "Content-Type": "application/json"
       },
       body: JSON.stringify(newMessageObj)
+    }).then(() => {
+      const activeUserId = sessionStorage.getItem("activeUser");
+      printToDOM(activeUserId);
     });
   }
 };
